fix(client): 프로필 페이지 전부 ssr 제거한 뒤 스켈레톤 처리
