+++ apps/client/components/Profile/edit/index.ts
@@ -1,1 +1,2 @@
 export { default as UserInfo } from "./UserInfo";
+export { default as RegisterButton } from "./RegisterButton";
