+++ apps/client/components/Profile/profile/index.tsx
@@ -1,0 +1,2 @@
+export { default as Info } from "./Info";
+export { default as EditButton } from "./EditButton";
