+++ apps/client/server/profile/queries.ts
@@ -14,8 +14,6 @@ export const useFetchUserInfo = (
     "onSuccess"
   >
 ) => {
-  // Suspense 로 감싸주지 못하므로, 단언을 해줍니다.
-
   return useSuspenseQuery(Keys.user(), getUserInfo, options);
 };
 
@@ -25,7 +23,6 @@ export const useFetchUserPosts = (
     "onSuccess"
   >
 ) => {
-  // Suspense 로 감싸주지 못하므로, 단언을 해줍니다.
   return useSuspenseQuery(Keys.posts(), getUserPosts, options);
 };
